Enable users of rspec to include SimpleBdd by:
require 'simple_bdd/rspec'
