+++ lib/simple_bdd.rb
@@ -9,17 +9,4 @@ module SimpleBdd
     alias_method method.downcase, method
   end
 
-  if defined?(::RSpec)
-    def self.included(base)
-      base.instance_eval do
-        def feature(message, &block)
-          describe message, &block
-        end
-
-        def scenario(message, &block)
-          it message, &block
-        end
-      end
-    end
-  end
 end
